+++ vaulted-for-reference/create_packages.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-./create_chromium_package.sh
-echo
-./create_firefox_package.sh